Implement LastTouchedBy
Add lastTouchedByCount to Ball.cs
Add Queue<PlayerController> lastTouched by to Ball.cs
Add SetLastTouchedBy(PlayerController palyer) to Ball.cs
Call SetLastTouchedBy from GameManager on BallShieldCollision

Keeps track of the last player to hit the ball with the shield. These players are stored in a queue to allow more backtracking so that a player doesn't "kill" themself in the stats.