Add Player Kill Tracking
Add OnBallGoalCollision to ball.cs, deactivates the ball and clears the lastTouchedBy list.
Fix SetLastTouchedBy to check for empty lastTouchedBy
Change lastTouchedBy from Queue to LinkedList
Add GetLastTouchedBy with optional param to exclude a player from the search.
Modify BallPlayerCollision in GameManager to Register Player Kills
Refactor KillPlayer to PlayerDead to prevent confusion between registering kills a player gets and the call to tell a player they are dead.
Add a list of players killed in the playercontroller with RegisterKill function to add a kill to the list with some temp debugging.