Added velocity curve option
Gave the ball the ability to switch between linear and inverse-square slowing curves using the "Curve Slow" checkbox in the Inspector. The ball's slow speed when using Curve Slow can be adjusted using the Curve Modifier value so the Slow Rate value doesn't have to be changed (thus messing with the linear slow rate).